# swagger/swagger.yaml
consumes:
    - application/json
definitions:
    ClusterConfiguration:
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            GroupName:
                type: string
            ID:
                format: uint64
                type: integer
            KubernetesConfiguration:
                items:
                    format: uint8
                    type: integer
                type: array
            UpdatedAt:
                format: date-time
                type: string
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    DeletedAt:
        $ref: '#/definitions/NullTime'
...
    Group:
        description: Group domain object defining a group
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            Hostname:
                type: string
            Name:
                type: string
            UpdatedAt:
                format: date-time
                type: string
            Users:
                items:
                    $ref: '#/definitions/User'
                type: array
            clusterConfiguration:
                $ref: '#/definitions/ClusterConfiguration'
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    Instance:
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            DeployLog:
                type: string
            GroupName:
                type: string
            ID:
                format: uint64
                type: integer
            Name:
                type: string
            Preset:
                type: boolean
            PresetID:
                format: uint64
                type: integer
            StackName:
                type: string
            UpdatedAt:
                format: date-time
                type: string
            UserID:
                format: uint64
                type: integer
            optionalParameters:
                items:
                    $ref: '#/definitions/InstanceOptionalParameter'
                type: array
                x-go-name: OptionalParameters
            requiredParameters:
                items:
                    $ref: '#/definitions/InstanceRequiredParameter'
                type: array
                x-go-name: RequiredParameters
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
    InstanceOptionalParameter:
        properties:
            name:
                description: 'TODO: Rename StackOptionalParameterID to Name'
                type: string
                x-go-name: StackOptionalParameterID
            value:
                type: string
                x-go-name: Value
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
    InstanceRequiredParameter:
        properties:
            name:
                description: 'TODO: Rename StackRequiredParameterID to Name'
                type: string
                x-go-name: StackRequiredParameterID
            value:
                type: string
                x-go-name: Value
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    NullTime:
        description: |-
            NullTime implements the Scanner interface so
            it can be used as a scan destination, similar to NullString.
        properties:
            Time:
                format: date-time
                type: string
            Valid:
                type: boolean
        title: NullTime represents a time.Time that may be null.
        type: object
        x-go-package: database/sql
    RefreshTokenRequest:
        properties:
            refreshToken:
                type: string
                x-go-name: RefreshToken
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/user
...
    SignUpRequest:
        properties:
            email:
                type: string
                x-go-name: Email
            password:
                type: string
                x-go-name: Password
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/user
...
    Tokens:
        description: Tokens domain object defining user tokens
        properties:
            access_token:
                type: string
                x-go-name: AccessToken
            expires_in:
                format: uint64
                type: integer
                x-go-name: ExpiresIn
            refresh_token:
                type: string
                x-go-name: RefreshToken
            token_type:
                type: string
                x-go-name: TokenType
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/token
...
    UpdateInstanceRequest:
        properties:
            optionalParameters:
                items:
                    $ref: '#/definitions/InstanceOptionalParameter'
                type: array
                x-go-name: OptionalParameters
            requiredParameters:
                items:
                    $ref: '#/definitions/InstanceRequiredParameter'
                type: array
                x-go-name: RequiredParameters
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/instance
    User:
        description: User domain object defining a user
        properties:
            AdminGroups:
                items:
                    $ref: '#/definitions/Group'
                type: array
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            Email:
                type: string
            Groups:
                items:
                    $ref: '#/definitions/Group'
                type: array
            ID:
                format: uint64
                type: integer
            UpdatedAt:
                format: date-time
                type: string
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
info:
    contact:
        email: info@dhis2.org
        url: https://github.com/dhis2-sre/im-manager
    description: Instance Manager
    license:
        name: TODO
    title: Instance Manager
    version: 0.1.0
paths:
...
    /instances/{id}:
        delete:
            description: Delete an instance by id
            operationId: deleteInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "202":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Delete instance
        get:
            description: Find an instance by id
            operationId: findById
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find instance
        put:
            description: Update an instance...
            operationId: updateInstance
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: Update instance request body parameter
                  in: body
                  name: Payload
                  required: true
                  schema:
                    $ref: '#/definitions/UpdateInstanceRequest'
            responses:
                "204":
                    description: Instance
                    schema:
                        $ref: '#/definitions/Instance'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Update instance
...
    /me:
        get:
            description: Current user details
            operationId: me
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: User details
...
    /refresh:
        post:
            description: Refresh user tokens
            operationId: refreshToken
            parameters:
                - description: Refresh token request body parameter
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/RefreshTokenRequest'
            responses:
                "201":
                    $ref: '#/responses/Tokens'
                "400":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            summary: Refresh tokens
...
    /tokens:
        post:
            description: Sign in... And get tokens
            operationId: signIn
            responses:
                "201":
                    $ref: '#/responses/Tokens'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - basicAuth: []
            summary: Sign in
    /users:
        delete:
            description: Sign out user... The authentication is done using oauth and JWT. A JWT can't easily be invalidated so even after calling this endpoint a user can still sign in assuming the JWT isn't expired. However, the token can't be refreshed using the refresh token supplied upon signin
            operationId: signOut
            responses:
                "200":
                    description: ""
                "401":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Sign out
        post:
            description: Sign up a user. This endpoint is publicly accessible and therefor anyone can sign up. However, before being able to perform any actions, users needs to be a member of a group. And only administrators can add users to groups.
            operationId: signUp
            parameters:
                - description: SignUp request body parameter
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/SignUpRequest'
            responses:
                "201":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            summary: SignUp user
    /users/{id}:
        get:
            description: Find a user by its id
            operationId: findUserById
            parameters:
                - format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find user
produces:
    - application/json
responses:
...
    Error:
        description: ""
...
    Tokens:
        description: ""
        schema:
            $ref: '#/definitions/Tokens'
securityDefinitions:
    oauth2:
        flow: password
        tokenUrl: /tokens
        type: oauth2
swagger: "2.0"